+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Test and beta releases will have experimental functions enabled by default.
 
 * Fixed handling for automated export, thanks @[jpeg729](https://github.com/jpeg729)
 * Added function name to on demand restricting dialog
+* Move database back to /data/xprivacy
 
 [Open issues](https://github.com/M66B/XPrivacy/issues?state=open)